+++ examples/components/wizard-view/wizard-view.component.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button } from 'react-bootstrap';
+import { Button, Checkbox } from 'react-bootstrap';
 
 import { Wizard } from '../../../src/index';
 
@@ -20,32 +20,42 @@ export default class WizardView extends React.Component {
     super();
     this.state = {
       showWizard: false,
+      disableSave: false,
+      disableClose: false,
     };
 
-    this.steps = [{
-      id: '1',
-      name: 'Wizard page 1',
-      component: this.getContent('Page 1'),
-    }, {
-      id: '2',
-      name: 'Wizard page 2',
-      component: this.getContent('Page 2'),
-    }, {
-      id: '3',
-      name: 'Wizard page 3',
-      component: this.getContent('Page 3'),
-    }, {
-      id: '4',
-      name: 'Wizard page 4',
-      component: this.getContent('Page 4'),
-      isValid: false,
-    }];
+    this.steps = [
+      {
+        id: '1',
+        hasRequiredProps: true,
+        name: 'Wizard page 1',
+        component: this.getContent('Page 1'),
+      },
+      {
+        id: '2',
+        hasRequiredProps: true,
+        hasRequiredPropsErrors: true,
+        name: 'Wizard page 2',
+        component: this.getContent('Page 2'),
+      },
+      {
+        id: '3',
+        name: 'Wizard page 3',
+        component: this.getContent('Page 3'),
+      },
+      {
+        id: '4',
+        name: 'Wizard page 4',
+        component: this.getContent('Page 4'),
+        isValid: false,
+      },
+    ];
   }
 
   getContent = (text) => {
     const content = (
       <div style={contentStyle}>
-        { text }
+        {text}
       </div>
     );
 
@@ -71,6 +81,18 @@ export default class WizardView extends React.Component {
     });
   }
 
+  toggleDisableSave = () => {
+    this.setState({
+      disableSave: !this.state.disableSave,
+    });
+  }
+
+  toggleDisableClose = () => {
+    this.setState({
+      disableClose: !this.state.disableClose,
+    });
+  }
+
   render() {
     return (
       <div className="oc-content" style={{ height: '100%' }}>
@@ -79,13 +101,30 @@ export default class WizardView extends React.Component {
             <Wizard
               save={this.saveWizard}
               cancel={this.cancelWizard}
+              disableSave={this.state.disableSave}
+              disableCancel={this.state.disableClose}
               steps={this.steps}
-              localizationTexts={{ save: 'Save', cancel: 'Cancel' }}
+              localizationTexts={{ save: 'Save', cancel: 'Close' }}
               showPageIndicator={false}
             />
-          : <Button onClick={this.showWizard}>
-              Start wizard...
-            </Button>}
+          :
+            <div>
+              <Button onClick={this.showWizard}>
+                Start wizard...
+                </Button>
+              <Checkbox
+                checked={this.state.disableSave}
+                onChange={this.toggleDisableSave}
+              >
+                Disable save button
+                </Checkbox>
+              <Checkbox
+                checked={this.state.disableClose}
+                onChange={this.toggleDisableClose}
+              >
+                Disable close button
+                </Checkbox>
+            </div>}
       </div>
     );
   }