+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added props `disableSave` and `disableCancel` to disable wizard buttons.
 * Default button titles in `Wizard` set to Save and Close
 * Only the active step is highlighted in the navbar of `Wizard` component.
+* Added ConfirmDialog component.
 
 ## 4.2.5
 * Add `hasRequiredProps` and `hasRequiredPropsErrors` and wrap `name` as a label in `Wizard`